drawing boids with OPENRNDR!

+++ src/main/kotlin/Boid.kt
@@ -2,20 +2,20 @@ import java.util.UUID
 import kotlin.random.Random
 
 data class Boid(
-    val x: Int,
-    val y: Int,
+    val x: Double,
+    val y: Double,
     val heading: Double,
-    val speed: Int,
+    val speed: Double,
     val id: UUID = UUID.randomUUID()
 ) {
     companion object {
         fun randomBoid() = Boid(
-            x = Random.nextInt(Config.SCREEN_SIZE_X.value.inc()),
-            y = Random.nextInt(Config.SCREEN_SIZE_Y.value.inc()),
+            x = Random.nextDouble(Config.SCREEN_SIZE_X.value.inc().toDouble()),
+            y = Random.nextDouble(Config.SCREEN_SIZE_Y.value.inc().toDouble()),
             heading = Random.nextDouble(Constants.TWO_PI),
-            speed = Random.nextInt(
-                Config.BOID_MIN_SPEED.value,
-                Config.BOID_MAX_SPEED.value.inc())
+            speed = Random.nextDouble(
+                Config.BOID_MIN_SPEED.value.toDouble(),
+                Config.BOID_MAX_SPEED.value.inc().toDouble())
         )
     }
 }
